app/backend/requirements-phase3.txt
# ============================================
# Fase 3: Backend Flask + MySQL
# Sistema de Controle de Acesso - CEU Tres Pontes
# ============================================

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0
# mysqlclient==2.2.0  # Removido - causa problemas no Windows (precisa compilar C)
# PyMySQL funciona perfeitamente como driver MySQL para SQLAlchemy

# MQTT (da Fase 2)
paho-mqtt==1.6.1

# Validation & Serialization
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Environment & Configuration
python-dotenv==1.0.0

# Date & Time
python-dateutil==2.8.2

# Utilities
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
Faker==20.1.0

# API Documentation
flask-swagger-ui==4.11.1
apispec==6.3.1
apispec-webframeworks==0.5.2

# Task Queue (RabbitMQ) - Opcional para Fase 3
# celery==5.3.4
# kombu==5.3.4

# Production Server (Linux/Unix apenas)
# gunicorn==21.2.0  # Não funciona no Windows - Use waitress ou o servidor dev do Flask
waitress==3.0.2  # Servidor WSGI para Windows
